feat(blog): setup apollo
