+++ src/pages/index.js
@@ -1,3 +1,4 @@
-export * from './home';
-export * from './not-found';
-export * from './wordle';
+export * from "./blog";
+export * from "./home";
+export * from "./not-found";
+export * from "./wordle";
